05/11/2021 S

+++ ft_convert.c
@@ -6,7 +6,7 @@
 /*   By: kcatrix <kcatrix@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/05 09:13:26 by kcatrix           #+#    #+#             */
-/*   Updated: 2021/11/05 15:20:21 by kcatrix          ###   ########.fr       */
+/*   Updated: 2021/11/05 16:38:56 by kcatrix          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	ft_convert(char arg, void *param)
 	if ((int)arg == 'p')
 		write(1, "p", 1);
 	if ((int)arg == 'd')
-		i += ft_putnbr((int)param);
+		i += ft_putnbr((long)param);
 	if ((int)arg == 'i')
 		write(1, "i", 1);
 	if ((int)arg == 'u')
